fix: code review fixes for story 3-4-environment-variable-substitution-utility
- Move InMemoryToolCatalogRepository to tests/helpers.py (eliminate cross-test coupling)
- Relocate AC5 test from test_env.py to test_tool_catalog.py (correct test placement)
- Strengthen AC5 assertion from 'in' to exact equality
- Add edge case tests: empty env var value, partial resolution failure

Relates to #21
