+++ tests/helpers.py
@@ -3,8 +3,42 @@
 import builtins
 
 from akgentic.catalog.models.agent import AgentEntry
+from akgentic.catalog.models.queries import ToolQuery
+from akgentic.catalog.models.tool import ToolEntry
+from akgentic.catalog.repositories.base import ToolCatalogRepository
 
 _list = builtins.list
+
+
+class InMemoryToolCatalogRepository(ToolCatalogRepository):
+    """Simple in-memory repository for testing service logic."""
+
+    def __init__(self) -> None:
+        self._entries: dict[str, ToolEntry] = {}
+
+    def create(self, tool_entry: ToolEntry) -> str:
+        self._entries[tool_entry.id] = tool_entry
+        return tool_entry.id
+
+    def get(self, id: str) -> ToolEntry | None:
+        return self._entries.get(id)
+
+    def list(self) -> _list[ToolEntry]:
+        return _list(self._entries.values())
+
+    def search(self, query: ToolQuery) -> _list[ToolEntry]:
+        results = self.list()
+        if query.id is not None:
+            results = [e for e in results if e.id == query.id]
+        if query.tool_class is not None:
+            results = [e for e in results if e.tool_class == query.tool_class]
+        return results
+
+    def update(self, id: str, tool_entry: ToolEntry) -> None:
+        self._entries[id] = tool_entry
+
+    def delete(self, id: str) -> None:
+        del self._entries[id]
 
 
 class MockAgentCatalogRepository:
